requirements.txt
crewai
crewai-tools
python-dotenv
langchain-groq  # free & fast (or langchain-openai if you prefer)
pydantic